# algorithm/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_createAnalysisJob": {
        "properties": {
          "calibration_file": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional UTF-8 dual-IMU CAL/validation CSV for raw_dual_imu input.",
            "title": "Calibration File"
          },
          "data_file": {
            "description": "UTF-8 CSV containing joint-state or dual-IMU task samples.",
            "format": "binary",
            "title": "Data File",
            "type": "string"
          },
          "mechanical_file": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional UTF-8 raw FSR, calibrated pressure, or operator-state CSV.",
            "title": "Mechanical File"
          },
          "metadata": {
            "description": "JSON metadata matching SheWrist Backend API v1.0.",
            "title": "Metadata",
            "type": "string"
          }
        },
        "required": [
          "metadata",
          "data_file"
        ],
        "title": "Body_createAnalysisJob",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "JobResponse": {
        "properties": {
          "analysis_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Status"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "progress_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Pct",
            "type": "integer"
          },
          "result_url": {
            "title": "Result Url",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "job_id",
          "session_id",
          "status",
          "stage",
          "progress_pct",
          "created_at",
          "updated_at",
          "status_url",
          "result_url"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "SessionResult": {
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "algorithm_release": {
            "title": "Algorithm Release",
            "type": "string"
          },
          "analysis_status": {
            "title": "Analysis Status",
            "type": "string"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "calibration": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration"
          },
          "channels": {
            "$ref": "#/components/schemas/Channels"
          },
          "control_policy": {
            "$ref": "#/components/schemas/ControlPolicy"
          },
          "data_quality": {
            "$ref": "#/components/schemas/DataQuality"
          },
          "evidence_limits": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Limits",
            "type": "array"
          },
          "evidence_type": {
            "title": "Evidence Type",
            "type": "string"
          },
          "explanation": {
            "$ref": "#/components/schemas/ExplanationSummary"
          },
          "fsr_proxy": {
            "$ref": "#/components/schemas/FsrProxySummary"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "metrics": {
            "$ref": "#/components/schemas/Metrics"
          },
          "ml_shadow": {
            "$ref": "#/components/schemas/MlShadowSummary"
          },
          "rejection_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Rejection Reasons",
            "type": "array"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "sensor_installation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SensorInstallation"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trial_condition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrialCondition"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "job_id",
          "session_id",
          "status",
          "analysis_status",
          "rejection_reasons",
          "evidence_type",
          "algorithm_release",
          "channels",
          "data_quality",
          "fsr_proxy",
          "metrics",
          "alerts",
          "ml_shadow",
          "control_policy",
          "explanation",
          "artifacts",
          "warnings",
          "evidence_limits"
        ],
        "title": "SessionResult",
        "type": "object"
      },
      "TimelineResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineRow"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "session_id",
          "offset",
          "limit",
          "total",
          "items"
        ],
        "title": "TimelineResponse",
        "type": "object"
      },
...
      "TokensResponse": {
        "properties": {
          "operating_mode": {
            "title": "Operating Mode",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "tokens": {
            "items": {
              "$ref": "#/components/schemas/InertialTokenResponse"
            },
            "title": "Tokens",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "operating_mode",
          "tokens"
        ],
        "title": "TokensResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Asynchronous offline wrist-exposure analysis. The deterministic exposure engine is the only alert path; the CNN-HMM and optional explanation provider have no control authority.",
    "title": "SheWrist Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analysis-jobs": {
      "post": {
        "operationId": "createAnalysisJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createAnalysisJob"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request or unsupported input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Job, session, or artifact not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Idempotency, session, or result-state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upload too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or analysis failure"
          }
        },
        "summary": "Create an asynchronous offline-analysis job",
        "tags": [
          "analysis jobs"
        ]
      }
    },
    "/api/v1/analysis-jobs/{job_id}": {
      "get": {
        "operationId": "getAnalysisJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request or unsupported input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Job, session, or artifact not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Idempotency, session, or result-state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upload too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or analysis failure"
          }
        },
        "summary": "Get asynchronous job status",
        "tags": [
          "analysis jobs"
        ]
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "operationId": "getSessionAnalysis",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request or unsupported input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Job, session, or artifact not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Idempotency, session, or result-state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upload too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or analysis failure"
          }
        },
        "summary": "Get the normalized analysis result",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/artifacts/{name}": {
      "get": {
        "operationId": "downloadSessionArtifact",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request or unsupported input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Job, session, or artifact not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Idempotency, session, or result-state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upload too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or analysis failure"
          }
        },
        "summary": "Download an allow-listed auditable artifact",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/timeline": {
      "get": {
        "operationId": "getSessionTimeline",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request or unsupported input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Job, session, or artifact not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Idempotency, session, or result-state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upload too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or analysis failure"
          }
        },
        "summary": "Get a paginated chart-ready timeline",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/tokens": {
      "get": {
        "operationId": "getSessionTokens",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request or unsupported input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Job, session, or artifact not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Idempotency, session, or result-state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upload too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation or analysis failure"
          }
        },
        "summary": "Get non-controlling CNN-HMM inertial tokens",
        "tags": [
          "sessions"
        ]
      }
    },
...
  }
}